Ninjamaster: Split out admin view object.
Fixes #457.
