+++ deploy/lib/control/NinjamasterController.php
@@ -2,9 +2,9 @@
 namespace NinjaWars\core\control;
 
 require_once(LIB_ROOT."data/NpcFactory.php");
+require_once(LIB_ROOT."data/AdminViews.php");
 require_once(ROOT.'core/data/AccountFactory.php');
 require_once(LIB_ROOT."data/Npc.php");
-require_once(LIB_ROOT."control/lib_inventory.php");
 require_once(LIB_ROOT."control/lib_player_list.php");
 require_once(LIB_ROOT."control/lib_player.php"); // For player tags
 
@@ -12,71 +12,16 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 use \Player;
 use \NpcFactory;
 use \AccountFactory;
+use \NinjaWars\core\data\AdminViews;
 
 
-class AdminViews{
-
-    public static function high_rollers(){
-        // Select first few max kills from players.
-        // Max turns.
-        // Max gold.
-        // Max kills
-        // etc.
-        $res = array();
-        $res['gold'] = query_array('select player_id, uname, gold from players order by gold desc limit 10');
-        $res['turns'] = query_array('select player_id, uname, turns from players order by turns desc limit 10');
-        $res['kills'] = query_array('select player_id, uname, kills from players order by kills desc limit 10');
-        $res['health'] = query_array('select player_id, uname, health from players order by health desc limit 10');
-        $res['ki'] = query_array('select player_id, uname, ki from players order by ki desc limit 10');
-        return $res;
-    }
-
-    public static function duped_ips(){
-        $host= gethostname();
-        $server_ip = gethostbyname($host);
-        // Get name, id, and ip from players, grouped by ip matches
-        return query('select uname, player_id, days, last_ip from players left join account_players on player_id = _player_id
-            left join accounts on _account_id = account_id where uname is not null 
-            and last_ip in 
-            (SELECT last_ip FROM accounts 
-                WHERE (operational = true and confirmed = 1) 
-                    and (last_ip != \'\' and last_ip != \'127.0.0.1\' and last_ip != :server_ip) 
-                GROUP  BY last_ip HAVING count(*) > 1 ORDER BY count(*) DESC limit 30)
-             order by last_ip, days ASC limit 300',
-             [':server_ip'=>$server_ip]);
-    }
-
-
-    // Reformat the character info sets.
-    public static function split_char_infos($ids){
-        if(is_numeric($ids)){
-            $ids = [$ids]; // Wrap it in an array.
-        } else { // Get the info for multiple ninjas
-            $res = array();
-            $ids = explode(',', $ids);
-        }
-        $first = true;
-        foreach($ids as $id){
-            $res[$id] = char_info($id, $admin_info=true);
-            $res[$id]['first'] = $first;
-            unset($res[$id]['messages']); // Exclude the messages for length reasons.
-            unset($res[$id]['description']); // Ditto
-            $first = false;
-        }
-        return $res;
-    }
-
-    public static function char_inventory($char_id){
-        return inventory_counts($char_id);
-    }
-}
 
 /**
  * The ninjamaster/admin info
  */
 class NinjamasterController {
     const ALIVE = false;
-    const PRIV  = true;
+    const PRIV  = false;
     protected $char_id = null;
 
     public function __construct(){
